SAK-22468 - Configure CKEditor to use user language
This patch uses the newly added sakai.locale information to load the
user's language in the CKEditor. It specifies defaultLanguage = 'en'
because CKEditor does not launch properly when the default is specified
as a language that is not supported. Within Sakai, the user's locale
already reflects the user preference, defaulting to the system locale.

This choice is a trade-off between the risk of the system locale not
having a CKEditor translation (thereby breaking all editor instances
unless a user switched to a supported locale) and the case where a
non-English system locale is in place and a user selects an unsupported
locale (failing back to English, rather than the system language).

This is certainly a valid point to debate. It is possible to expose the
system locale and use it as the default, but if the CKEditor translation
were not available, the implementor would have to modify
ckeditor.launch.js to use a supported locale or add a language pack to
CKEditor.

If it turns out that CKEditor has all of our translations covered,
setting defaultLanguage to the Sakai default language would be the best
course of action.



git-svn-id: https://source.sakaiproject.org/svn/reference/trunk@111802 66ffb92e-73f9-0310-93c1-f5514f145a0a